# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub William API",
    "description": "Control-plane and gateway API for Hub William agent pools",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/agent-pool-requests/{request_id}/decision": {
      "post": {
        "tags": ["agent pools"],
        "operationId": "decide_request",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "description": "Join request identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideAgentPoolJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Join request decided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPoolJoinRequest"
                }
              }
            }
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent-pools/{connection_id}/members": {
      "post": {
        "tags": ["agent pools"],
        "operationId": "invite_member",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connected account identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAgentPoolMember"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member invited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPoolPerson"
                }
              }
            }
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invite rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-pools/{connection_id}/members/{username}": {
      "delete": {
        "tags": ["agent pools"],
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connected account identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "username",
            "in": "path",
            "description": "Bew Harness username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed"
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Membership not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-pools/{connection_id}/requests": {
      "post": {
        "tags": ["agent pools"],
        "operationId": "create_request",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connected account identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentPoolJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Join request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPoolJoinRequest"
                }
              }
            }
          },
          "401": {
            "description": "Login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentPoolJoinRequest": {
        "type": "object",
        "required": [
          "avatar_label",
          "id",
          "reason",
          "status",
          "telegram",
          "username"
        ],
        "properties": {
          "avatar_label": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentPoolRequestStatus"
          },
          "telegram": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "AgentPoolPerson": {
        "type": "object",
        "required": [
          "avatar_label",
          "joined_at",
          "share",
          "usage_available_percent",
          "username"
        ],
        "properties": {
          "avatar_label": {
            "type": "string"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "share": {
            "$ref": "#/components/schemas/AgentPoolShareEvidence"
          },
          "usage_available_percent": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "AgentPoolRequestStatus": {
        "type": "string",
        "enum": ["pending", "accepted", "rejected"]
      },
      "AgentPoolShareEvidence": {
        "type": "object",
        "required": [
          "available_percent",
          "member_count",
          "pool_cached_tokens",
          "pool_input_tokens",
          "pool_output_tokens",
          "pool_units",
          "user_cached_tokens",
          "user_input_tokens",
          "user_output_tokens",
          "user_units"
        ],
        "properties": {
          "available_percent": {
            "type": "integer",
            "format": "int32"
          },
          "budget_units": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "cap_units": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "fail_open_reason": {
            "type": ["string", "null"]
          },
          "member_count": {
            "type": "integer",
            "format": "int32"
          },
          "pool_cached_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "pool_input_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "pool_output_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "pool_units": {
            "type": "integer",
            "format": "int64"
          },
          "provider_used_percent": {
            "type": ["number", "null"],
            "format": "double"
          },
          "remaining_units": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "user_cached_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "user_input_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "user_output_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "user_units": {
            "type": "integer",
            "format": "int64"
          },
          "window_label": {
            "type": ["string", "null"]
          }
        }
      },
...
      "CreateAgentPoolJoinRequest": {
        "type": "object",
        "required": ["reason", "telegram"],
        "properties": {
          "reason": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          }
        }
      },
...
      "DecideAgentPoolJoinRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AgentPoolRequestStatus"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "InviteAgentPoolMember": {
        "type": "object",
        "required": ["username"],
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}